requirements.txt
telethon==1.34.0
rich==13.7.1
cryptography==42.0.5
python-dotenv==1.0.1